requirements.txt
# Core data pipeline
google-cloud-bigquery>=3.0.0
psycopg2-binary>=2.9
python-dateutil>=2.8

# Testing & code quality
pytest>=7.0
pytest-cov>=3.0
detect-secrets>=1.1.0

# Data science & notebooks
jupyter>=1.0
pandas>=1.3
matplotlib>=3.4
seaborn>=0.11
numpy>=1.21

# Utilities
python-dotenv>=0.19
click>=8.0
tqdm>=4.60

# Optional: API & dashboards
flask>=2.0
sqlalchemy>=1.4
psycopg2>=2.9

# Logging & monitoring
python-json-logger>=2.0
